% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{deseqRes2Tibble}
\alias{deseqRes2Tibble}
\title{Convert DESeq2 results table into a tibble}
\usage{
deseqRes2Tibble(res)
}
\value{
results table in tibble format
}
\description{
Convert DESeq2 results table into a tibble
}
