% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{factor2Char}
\alias{factor2Char}
\title{Convert a factor to a character vector, while preserving the names}
\usage{
factor2Char(f)
}
\arguments{
\item{f}{the factor to convert}
}
\value{
a named character vector
}
\description{
Convert a factor to a character vector, while preserving the names
}
