% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{factorBreakdown.2}
\alias{factorBreakdown.2}
\title{Break down a factor returning the names of the elements
in each level as character vectors in a list}
\usage{
factorBreakdown.2(f)
}
\arguments{
\item{f}{a factor to breakdown}
}
\value{
a list of factor levels with the names of elemements in them
}
\description{
Break down a factor returning the names of the elements
in each level as character vectors in a list
}
