% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getDataWithCovarianceMatrix}
\alias{getDataWithCovarianceMatrix}
\title{Generates data with covariance structure specified by the given covariance matrix}
\usage{
getDataWithCovarianceMatrix(covM, nobs)
}
\arguments{
\item{covM}{a posititive definite covariance matrix}

\item{nobs}{the number of observations requestest}
}
\value{
data.frame object with the observations as rows and the variables as columns
}
\description{
Generates the requested number of observations that follow the given
covariance structure specified by the given matrix
}
