% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFactorFromP1rds}
\alias{getFactorFromP1rds}
\title{Get the colors of a pagoda1 serialised app object as a factor}
\usage{
getFactorFromP1rds(filename, colcol.name)
}
\arguments{
\item{filename}{the rds file to load}

\item{colcol.name}{the name of the colcol item to retrive}
}
\value{
a named factor
}
\description{
get the colors of a pagoda1 serialised app object as a factor from the app$results$colcol list
}
