% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getNamedDirectoryListing}
\alias{getNamedDirectoryListing}
\title{Like list.files but return the full path and the vector entries
are named only after the basename}
\usage{
getNamedDirectoryListing(path)
}
\arguments{
\item{path}{location to list}
}
\description{
Like list.files but return the full path and the vector entries
are named only after the basename
}
