% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getSlurmCpus}
\alias{getSlurmCpus}
\title{Get the number of processors allocated to the current job by
the Slurm scheduler from the environment. Return 1 otherwise}
\usage{
getSlurmCpus()
}
\description{
Get the number of processors allocated to the current job by
the Slurm scheduler from the environment. Return 1 otherwise
}
