% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{headtail}
\alias{headtail}
\title{Head and tail of a table combined}
\usage{
headtail(x, n = 6)
}
\arguments{
\item{x}{the table object}

\item{n}{a single integer}
}
\value{
a table made of the n first elemets of x and n last elements of x
}
\description{
Head and tail of a table combined
}
