% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ll2a}
\alias{ll2a}
\title{Convert list of lists of numbers to array}
\usage{
ll2a(inputList)
}
\arguments{
\item{inputList}{a list of lists of numbers}
}
\value{
an array where the top-level lists are rows and low-level lists are columns
}
\description{
Converts lists of lists of numbers to a numeric array fast
this has been optimised against other possible implementations and
been found to the the fastest
}
\examples{
mylist <- list()
mylist_ <- list()
for(i in 1:10) {
  for(j in 1:100) {
    mylist[[j]] <- i*j
  }
  mylist_[[i]] <- mylist
}
str(mylist_)
r3 <- ll2a.3(mylist_)
}
