% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load.hca.matrix}
\alias{load.hca.matrix}
\title{Load a sparse matrix from an HDF5 file ensuring that the
second dimension is sorted}
\usage{
load.hca.matrix(path)
}
\arguments{
\item{path}{location of the HDF files}
}
\value{
a Matrix sparse matrix
}
\description{
Load a sparse matrix from an HDF5 file ensuring that the
second dimension is sorted
}
