% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load.image.fast}
\alias{load.image.fast}
\title{Loads an image generated with save.image.fast()}
\usage{
load.image.fast(filename, tmpfile = NULL, verbose = F,
  envir = parent.frame())
}
\arguments{
\item{filename}{filename to load}

\item{tmpfile}{temporary file to use, use of ramdisk file will accelerate loading}

\item{verbose}{logical verbosity level}

\item{envir}{enviroment in which to load the data, by default the calling environment}
}
\description{
loads an image generated with the save.image.fast function. This function
requires that the system has the lbunzip2 command installed
}
