% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loadKleinTables}
\alias{loadKleinTables}
\title{Load TSV tables generated by the Klein pipeline as sparse matrices}
\usage{
loadKleinTables(filepaths, min.common.genes = 1000, prefix.sep = "_")
}
\arguments{
\item{filepaths}{named vector of file paths}

\item{min.common.genes}{the minimum number of commong genes to accept}

\item{prefix.sep}{separator for prefixing cells with sample id}
}
\value{
list of sparse matrices
}
\description{
Load TSV tables generated by the Klein pipeline as sparse matrices
}
