% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.doublets.two.groups}
\alias{make.doublets.two.groups}
\title{Makes an array with artificial doublets}
\usage{
make.doublets.two.groups(cm, group1, group2, mix.ratio.min, mix.ratio.max, n,
  mc.cores = 1)
}
\arguments{
\item{cm}{count matrix, rows are genes, cols are cells}

\item{n}{number of cells to generate}
}
\value{
a new array with artificial doublets
}
\description{
makes an array with artificial doublets
}
\examples{

group1 <- colnames(cm)[sample(1:length(colnames(cm)),1000)]
group2 <- colnames(cm)[sample(1:length(colnames(cm)),1000)]
int <- intersect(group1,group2)
length(int)
group1 <- setdiff(group1, int)
group2 <- setdiff(group2, int)
doublets <- make.doublets(cm=cm, group1=group1, group2=group2, mix.ratio.min=0.4,mix.ratio.max=0.6, n=100, mc.cores=40)
}
