% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make.doublets.two.groups.subsample}
\alias{make.doublets.two.groups.subsample}
\title{Makes an array with artificial doublets by subsampling of different mix ratios}
\usage{
make.doublets.two.groups.subsample(cm, group1, group2, mix.ratio.min,
  mix.ratio.max, n, mc.cores = 1, oversample = 1.5)
}
\arguments{
\item{cm}{count matrix, rows are genes, cols are cells}

\item{group1}{group1 to sample from}

\item{group2}{group2 to sample from}

\item{mix.ratio.min}{minimum mixing ratio}

\item{mix.ratio.max}{maximum mixing ratio}

\item{n}{number of cells to generate}

\item{oversample}{oversample/undersampling oversampling of reads}
}
\value{
a new array with artificial doublets
}
\description{
makes an array with artificial doublets  by subsampling of different mix ratios
}
\examples{

group1 <- colnames(cm)[sample(1:length(colnames(cm)),1000)]
group2 <- colnames(cm)[sample(1:length(colnames(cm)),1000)]
int <- intersect(group1,group2)
length(int)
group1 <- setdiff(group1, int)
group2 <- setdiff(group2, int)
doublets <- make.doublets(cm=cm, group1=group1, group2=group2, mix.ratio.min=0.4,mix.ratio.max=0.6, n=100, mc.cores=40, oversample = 1.5)
}
