% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{mergeFactor}
\alias{mergeFactor}
\title{Merge a previously broken down factor}
\usage{
mergeFactor(bdf)
}
\arguments{
\item{bdf}{a broken down factor}
}
\value{
a factor
}
\description{
Merge a previously broken down factor
}
