% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mergeMatrices}
\alias{mergeMatrices}
\title{Merge expression matrices in a list, keeping only common genes}
\usage{
mergeMatrices(data.all)
}
\arguments{
\item{data.all}{a list of sparse matrices}
}
\value{
a single merged matrix
}
\description{
Merge expression matrices in a list, keeping only common genes
}
