% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{minStringDistance}
\alias{minStringDistance}
\title{Find minimum string distance strings in a vector}
\usage{
minStringDistance(strings)
}
\description{
Find the minimum string distance between strings in a vector,
this is useful for things like finding the number of allowable barcode
mismatches
}
