% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixCellsInMatrix.R
\name{mixCellsInMatrix}
\alias{mixCellsInMatrix}
\title{mix cells in a matrix with a common background profile generated from all the cell}
\usage{
mixCellsInMatrix(mat, bg.prop, n.cores = 1, verbose = FALSE)
}
\arguments{
\item{n.cores}{number of cores t use}

\item{verbose}{print out dots to show while processing}

\item{matrix}{of genes x cells}

\item{bg.prog}{proportion of read to come from the background}
}
\value{
object of class dgCMatrix
}
\description{
mix cells in a matrix with a common background profile generated from all the cell
}
