% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mixVectorsOnLogical}
\alias{mixVectorsOnLogical}
\title{Generates a new vector from two vectors by combining them according to a third}
\usage{
mixVectorsOnLogical(logical, vTrue, vFalse)
}
\arguments{
\item{logical}{vector to use to combine the vTrue and vFalse vectors}

\item{vTrue}{values to return if the logical is true}

\item{vFalse}{values to return if the logical is false}
}
\value{
a new vector with some values from vTrue and some for vFalse
}
\description{
generates a vector from two vectors depending of the logical value of
a third vector
}
