% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbHelpers.R
\docType{package}
\name{nbHelpers}
\alias{nbHelpers}
\alias{nbHelpers-package}
\title{nbHelpers}
\description{
A package with a collection of helper functions for bioinformatics analyses
}
\author{
Nikolas Barkas
}
