% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{plotFactorsPercent}
\alias{plotFactorsPercent}
\title{Generate a ggplot2 plot comparing two factors for a common set of datapoints}
\usage{
plotFactorsPercent(factorA, factorB, points = NULL,
  factorA.name = "factorA", factorB.name = "factorB")
}
\arguments{
\item{factorA}{the factor the distibution of which with respect to factorB will be assessed (y axis)}

\item{factorB}{the factor to breakdown factorA by}

\item{points}{the names of the elemets of the factorsA and factorB to include for the
purposed of the plot}
}
\description{
Generates a ggplot barplot that compares two names two nameed factors
for the specified points. The x axis corresponds to factor B and the y axis
factor A.
}
\examples{
fA <- rep(c('A','B','C','D'), each = 4);
fB <- rep(c('W','Y'), each = 8)
n <- paste0('n',1:16)
names(fA) <- n
names(fB) <- n
plotFactorsPercent(fA, fB, points = names(fA)[1:9])
}
