% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{preserve.state}
\alias{preserve.state}
\title{Save the session in the current working directory with a file name that
includes the time stamp and process id}
\usage{
preserve.state(prefix = "savepoint_")
}
\arguments{
\item{prefix}{the prefix to use,for the filename}
}
\description{
Save the session in the current working directory with a file name that
includes the time stamp and process id
}
