% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{randomPositiveDefiniteMatrix}
\alias{randomPositiveDefiniteMatrix}
\title{Generates a random positive definite matrix}
\usage{
randomPositiveDefiniteMatrix(n, ev = runif(n, 0, 10))
}
\arguments{
\item{n}{dimensions}

\item{ev}{optional eigenvalues}
}
\value{
a random positive definite matrix with requested eigenvalues
}
\description{
generates a random positive definite matrix with the specified
eigenvalues. Code from: https://stat.ethz.ch/pipermail/r-help/2008-February/153708.html
}
