% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read10xMatrix}
\alias{read10xMatrix}
\title{Read 10x matrix}
\usage{
read10xMatrix(path)
}
\arguments{
\item{path}{location of 10x output}
}
\value{
None
}
\description{
This function reads a 10x matrix from the specified directory.
}
\author{
Nikolas Barkas
}
