% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLoad.R
\name{readInDropMatrix}
\alias{readInDropMatrix}
\title{Reads in a data matrix from the in house indrop pipeline}
\usage{
readInDropMatrix(name, path)
}
\arguments{
\item{name}{prefix to give to cells}

\item{path}{the file path}
}
\value{
a sparse expression matrix
}
\description{
Reads in a data matrix from the in house indrop pipeline
}
