% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readKleinMatrix}
\alias{readKleinMatrix}
\title{Reads in an expression matrix as formatted by the Klein lab
pipeline}
\usage{
readKleinMatrix(path, prefix)
}
\arguments{
\item{path}{path of the file}

\item{prefix}{prefix to add to the cell names}
}
\value{
a sparse matrix
}
\description{
Reads in an expression matrix as formatted by the Klein lab
pipeline
}
