% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLoad.R
\name{readMultiple10XmatricesAsList}
\alias{readMultiple10XmatricesAsList}
\title{read multiple 10x matrices and return as a list}
\usage{
readMultiple10XmatricesAsList(pathList)
}
\arguments{
\item{matrices}{a names list of paths to the matrices (that can be read by read10XMatrix)}
}
\value{
a list
}
\description{
given a named list of paths of 10X matrices return a list of matrices
}
