% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readMultipleInDropMatrices}
\alias{readMultipleInDropMatrices}
\title{Reads in multiple indrop matrices}
\usage{
readMultipleInDropMatrices(file.names)
}
\arguments{
\item{file.names}{named vector of filenames, names will become prefixes}
}
\value{
list of matrices
}
\description{
Reads in multiple indrop matrices
}
