% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLoad.R
\name{readMultipleKleinMatrices}
\alias{readMultipleKleinMatrices}
\title{Reads in multiple klein matrices and returns them in a list}
\usage{
readMultipleKleinMatrices(file.names)
}
\arguments{
\item{file.names}{named list of the files to load, the names will become prefixes}
}
\value{
a list of sparse matrices
}
\description{
Reads in multiple klein matrices and returns them in a list
}
