% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{removeLevelsBySize}
\alias{removeLevelsBySize}
\title{Remove levels and elements from a factor that have fewer than min.size entries}
\usage{
removeLevelsBySize(f, min.size = 3)
}
\arguments{
\item{f}{factor}

\item{min.size}{minimum size levels to keep}
}
\value{
a new factor
}
\description{
Remove levels and elements from a factor that have fewer than min.size entries
}
