% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{removeLevelsFromFactor}
\alias{removeLevelsFromFactor}
\title{Remove all elements from specificed clusters from a factor
and drop the respective levels}
\usage{
removeLevelsFromFactor(f, cl.remove)
}
\arguments{
\item{f}{factor}

\item{cl.remove}{character vector f clusters to remove}
}
\value{
a new factor
}
\description{
Remove all elements from specificed clusters from a factor
and drop the respective levels
}
