% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorHelpers.R
\name{replaceClusterFactors}
\alias{replaceClusterFactors}
\title{Combine two named factors when the new factor contains a breaking up of the
a subset of the original clusters}
\usage{
replaceClusterFactors(originalFactor, newFactor, newPrefix = NULL)
}
\arguments{
\item{originalFactor}{the factor in which to breakup clusters, this is not modified}

\item{newFactor}{the new factor specifying new cluster identities for some of the elements}
}
\value{
a new factor with the above combination of the factos
}
\description{
Combine two named factors when the new factor contains a breaking up of the
a subset of the original clusters
}
