% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save.image.fast}
\alias{save.image.fast}
\title{save a R session image fast}
\usage{
save.image.fast(filename, tmpfile = NULL, verbose = FALSE)
}
\arguments{
\item{filename}{the file to save the image to}

\item{tmpfile}{optional temporary file name, using a ramdisk location will further accelerate saving}

\item{verbose}{logical verbosity level}
}
\description{
saves an R image of the current session much faster and with
better compression that the built-in save.image function. Requires system lbzip2
utility
}
