% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setDisplay}
\alias{setDisplay}
\title{Set the display parameter for X11 plotting}
\usage{
setDisplay(display.number = NULL, host = "localhost")
}
\arguments{
\item{display.number}{the display number to set}

\item{host}{the host to set, defaults to localhost
return NUll}
}
\description{
Set the display parameter for X11 plotting
}
