% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setNames}
\alias{setNames}
\title{Set the element of an objects as the names
Useful in combination with lapply}
\usage{
setNames(x)
}
\arguments{
\item{x}{the object to process}
}
\value{
the same object with the elements named after conversion to character
}
\description{
Set the element of an objects as the names
Useful in combination with lapply
}
