% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sparseMatrixDensity}
\alias{sparseMatrixDensity}
\title{Get the percent of entries that are filled in a sparse matrix}
\usage{
sparseMatrixDensity(mat)
}
\arguments{
\item{mat}{matrix of class dgCMatrix or dgRMatrix}
}
\value{
percent of entries that are non-zero
}
\description{
Get the percent of entries that are filled in a sparse matrix
}
