% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{strpart}
\alias{strpart}
\title{Get the specified substring of string delimited by a character}
\usage{
strpart(x, split, n)
}
\arguments{
\item{x}{the strings to breakdown now (can be vector)}

\item{split}{the character to breakdown by}

\item{integer}{specifying which substing to get}
}
\value{
extracted substrings
}
\description{
Get the specified substring of string delimited by a character
}
