% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSampleCountMatrix.R
\name{subSampleCountMatrix}
\alias{subSampleCountMatrix}
\title{Subsample a count matrix to a specific number of reads}
\usage{
subSampleCountMatrix(mat, final.count = 1e+07)
}
\arguments{
\item{mat}{sparse matrix of class dgCMatrix}

\item{final.count}{final number of reads the matrix should have}
}
\value{
a new dgCMatrix subsample to the requested number of reads
}
\description{
Subsample a count matrix to a specific number of reads
}
