% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subSampleCountMatrix.R
\name{subSampleCountMatrix2}
\alias{subSampleCountMatrix2}
\title{Subsample a count matrix to a specific number of reads, improved version
ensures that every read has the same probability of being dropped}
\usage{
subSampleCountMatrix2(mat, final.count = 1e+06, seed = NULL)
}
\arguments{
\item{mat}{sparse matrix of class dgCMatrix}

\item{final.count}{final number of reads the matrix should have}

\item{seed}{seed for random number generator}
}
\value{
a new dgCMatrix subsample to the requested number of reads
}
\description{
Subsample a count matrix to a specific number of reads, improved version
ensures that every read has the same probability of being dropped
}
