% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{symmsetdiff}
\alias{symmsetdiff}
\title{Get the symmetrical set difference between two sets (XOR)}
\usage{
symmsetdiff(a, b)
}
\arguments{
\item{a}{first set}

\item{b}{sencond set}
}
\value{
elements in one of the two sets but not in the intersection
}
\description{
Get the symmetrical set difference between two sets (XOR)
}
