% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tableCorners}
\alias{tableCorners}
\title{get NW, NE, SW and SE corners of a dataframe or matrix}
\usage{
tableCorners(vals, n = 4)
}
\arguments{
\item{vals}{the dataframe or matrix of interest}

\item{n}{how many rows/cols to return from each end}
}
\value{
the new matrix
}
\description{
return a new dataframe or matrix with the elements at the
four corners of the one provided
}
