% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{transpose_dgRMatrix}
\alias{transpose_dgRMatrix}
\title{Transpose a dgRMatrix and simultaneously convert it to dgCMatrix}
\usage{
transpose_dgRMatrix(inmat)
}
\arguments{
\item{inmat}{input matrix in dgRMatrix format}
}
\value{
A dgCMatrix that is the transposed dgRMatrix
}
\description{
Transpose a dgRMatrix and simultaneously convert it to dgCMatrix
}
