% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{uniqCombs}
\alias{uniqCombs}
\title{Get all unique combinations of elements in a list}
\usage{
uniqCombs(nms)
}
\arguments{
\item{nms}{elements to get combinations}
}
\value{
list of pairs of elements
}
\description{
Get all unique combinations of elements in a list
}
