% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vectorSeqOverlap}
\alias{vectorSeqOverlap}
\title{Get the number of commong elements among the top n elements
of two character vectors where n varies between a specified range}
\usage{
vectorSeqOverlap(itemsA, itemsB, start = 1, seq_end = NULL)
}
\arguments{
\item{itemsA}{character vector A}

\item{itemsB}{character vector B}

\item{start}{start of overlap sequence}

\item{seq_end}{end of overlap sequence,
if NULL up to the size of the smallest of the two lists (default: NULL)}
}
\value{
dataframe with two columns
}
\description{
Get the number of commong elements among the top n elements
of two character vectors where n varies between a specified range
}
