% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{vennCounts}
\alias{vennCounts}
\title{Given two vectors return the number of elements
only in a, in their intersection and only in b}
\usage{
vennCounts(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
a vector with (element count only in a, element count in intersection, element count only in b)
}
\description{
Given two vectors return the number of elements
only in a, in their intersection and only in b
}
