% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{writeListOfDFs}
\alias{writeListOfDFs}
\title{Write a list of data.frames as individual csv files}
\usage{
writeListOfDFs(x, output.dir)
}
\arguments{
\item{x}{named list of data.frames}

\item{output.dir}{the directory to save the files in}
}
\description{
Write a list of data.frames as individual csv files
}
