% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read10xmatrices}
\alias{read10xmatrices}
\title{read multiple 10x matrices into a single sparse array}
\usage{
read10xmatrices(paths, min.common.genes = 1000, common.genes = T,
  merge = T, prefix.cells = T, prefix.sep = "_")
}
\arguments{
\item{paths}{named vector of location of the data (readable by read10Xmatrix())}

\item{min.common.genes}{minimum number of common genes to allow}

\item{common.genes}{logical, subset all matrices to common genes, required for merge}

\item{merge}{logical, merge all the matrices to one, requires common.genes and prefix.cells}

\item{prefix.cells}{prefix all cells with the name of the respective path in paths}

\item{prefix.sep}{separator for prefix of cells}
}
\value{
a sparce matrix of the Matrix package that contains all the data prefixes by the corresponding sample name
}
\description{
given a named list of paths of 10X matrices return a single large matrix
with all the data and cell prefixed with the corresponding sample name
}
