% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLoad.R
\name{read10xMatrix}
\alias{read10xMatrix}
\title{Read 10x matrix}
\usage{
read10xMatrix(path, version = "V2", transcript.id = "SYMBOL")
}
\arguments{
\item{path}{location of 10x output}

\item{version}{version of 10x output to read, must be one of 'V2' or 'V3'}

\item{transcript.id}{transcript identifier to use, can be SYMBOL or ENSEMBL}
}
\value{
read matrix
}
\description{
This function reads a matrix generated by the 10x processing pipeline
from the specified directory and returns it. It aborts if one of the required
files in the specified directory do not exist.
}
