% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read10xMatrix}
\alias{read10xMatrix}
\title{Read 10x matrix}
\usage{
read10xMatrix(path)
}
\arguments{
\item{path}{location of 10x output}
}
\value{
read matrix
}
\description{
This function reads a matrix generated by the 10x processing pipeline
from the specified directory and returns it. It aborts if one of the required
files in the specified directory do not exist.
}
