% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{readMultiple10Xmatrices}
\alias{readMultiple10Xmatrices}
\title{read multiple 10x matrices into a single sparse array}
\usage{
readMultiple10Xmatrices(matrices, min.common.genes = 1000)
}
\arguments{
\item{matrices}{a names list of paths to the matrices (that can be read by read10XMatrix)}

\item{min.common.genes}{minimum number of common genes to allow}
}
\value{
a sparce matrix of the Matrix package that contains all the data prefixes by the corresponding sample name
}
\description{
given a named list of paths of 10X matrices return a single large matrix
with all the data and cell prefixed with the corresponding sample name
}
